Added correct logger and getWriter method to NativeIndexWriter
